// sdk/daml-lf/interpreter/src/test/assemblyscript/hello-world/protobuf/com/digitalasset/daml/lf/value/Value/List.ts
// Code generated by protoc-gen-as. DO NOT EDIT.
// Versions:
//   protoc-gen-as v1.3.0
//   protoc        v3.8.0

import { Writer, Reader, Protobuf } from "as-proto/assembly";
import { Value } from "../Value";

export class List {
  static encode(message: List, writer: Writer): void {
    const elements = message.elements;
    for (let i: i32 = 0; i < elements.length; ++i) {
      writer.uint32(10);
      writer.fork();
      Value.encode(elements[i], writer);
      writer.ldelim();
    }
  }

  static decode(reader: Reader, length: i32): List {
    const end: usize = length < 0 ? reader.end : reader.ptr + length;
    const message = new List();

    while (reader.ptr < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.elements.push(Value.decode(reader, reader.uint32()));
          break;

        default:
          reader.skipType(tag & 7);
          break;
      }
    }

    return message;
  }

  elements: Array<Value>;

  constructor(elements: Array<Value> = []) {
    this.elements = elements;
  }
}

export function encodeList(message: List): Uint8Array {
  return Protobuf.encode(message, List.encode);
}

export function decodeList(buffer: Uint8Array): List {
  return Protobuf.decode<List>(buffer, List.decode);
}