// sdk/daml-lf/interpreter/src/test/assemblyscript/hello-world/protobuf/com/digitalasset/daml/lf/value/Value/Map.ts
// Copyright (c) 2024 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
// SPDX-License-Identifier: Apache-2.0

// Code generated by protoc-gen-as. DO NOT EDIT.
// Versions:
//   protoc-gen-as v1.3.0
//   protoc        v3.8.0

import { Writer, Reader, Protobuf } from "as-proto/assembly";
import { Entry } from "./Map/Entry";

export class Map {
  static encode(message: Map, writer: Writer): void {
    const entries = message.entries;
    for (let i: i32 = 0; i < entries.length; ++i) {
      writer.uint32(10);
      writer.fork();
      Entry.encode(entries[i], writer);
      writer.ldelim();
    }
  }

  static decode(reader: Reader, length: i32): Map {
    const end: usize = length < 0 ? reader.end : reader.ptr + length;
    const message = new Map();

    while (reader.ptr < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.entries.push(Entry.decode(reader, reader.uint32()));
          break;

        default:
          reader.skipType(tag & 7);
          break;
      }
    }

    return message;
  }

  entries: Array<Entry>;

  constructor(entries: Array<Entry> = []) {
    this.entries = entries;
  }
}

export function encodeMap(message: Map): Uint8Array {
  return Protobuf.encode(message, Map.encode);
}

export function decodeMap(buffer: Uint8Array): Map {
  return Protobuf.decode<Map>(buffer, Map.decode);
}