// sdk/daml-lf/interpreter/src/test/assemblyscript/hello-world/protobuf/com/digitalasset/daml/lf/value/Value/Record.ts
// Code generated by protoc-gen-as. DO NOT EDIT.
// Versions:
//   protoc-gen-as v1.3.0
//   protoc        v3.8.0

import { Writer, Reader, Protobuf } from "as-proto/assembly";
import { Field } from "./Record/Field";

export class Record {
  static encode(message: Record, writer: Writer): void {
    const fields = message.fields;
    for (let i: i32 = 0; i < fields.length; ++i) {
      writer.uint32(10);
      writer.fork();
      Field.encode(fields[i], writer);
      writer.ldelim();
    }
  }

  static decode(reader: Reader, length: i32): Record {
    const end: usize = length < 0 ? reader.end : reader.ptr + length;
    const message = new Record();

    while (reader.ptr < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.fields.push(Field.decode(reader, reader.uint32()));
          break;

        default:
          reader.skipType(tag & 7);
          break;
      }
    }

    return message;
  }

  fields: Array<Field>;

  constructor(fields: Array<Field> = []) {
    this.fields = fields;
  }
}

export function encodeRecord(message: Record): Uint8Array {
  return Protobuf.encode(message, Record.encode);
}

export function decodeRecord(buffer: Uint8Array): Record {
  return Protobuf.decode<Record>(buffer, Record.decode);
}